edit tree example

+++ Example/SAB00700Front/DTOs/CategoryGridDTO.cs
@@ -2,7 +2,7 @@
 {
     public class CategoryGridDTO
     {
-        public int Id { get; set; }
+        public Int16 Id { get; set; }
         public string Name { get; set; }
     }
 }
